fix auto read not work

+++ book_kit/src/main/java/net/lzbook/kit/utils/OpenUDID.java
@@ -43,13 +43,17 @@ public class OpenUDID {
                 e.putString(PREF_KEY, _openUdid);
                 e.putLong(TIME, _time);
                 e.apply();
+
+                FileUtils.writeByteFile(filePath, MultiInputStreamHelper.encrypt(_openUdid.getBytes()));
             } else {
                 _openUdid = _keyInPref;
                 _time = _keyTime;
 
             }
 
-            FileUtils.writeByteFile(filePath, MultiInputStreamHelper.encrypt(_openUdid.getBytes()));
+            if (!new File(filePath).exists()) {
+                FileUtils.writeByteFile(filePath, MultiInputStreamHelper.encrypt(_openUdid.getBytes()));
+            }
             AppLog.d(TAG, "_openUdid= " + _openUdid);
         }
     }
